CI: Align Python version between RollPyTorch and Build Tests
The RollPyTorch action uses Python v3.11 (so does the Release build),
but the workflow that runs the CI tests uses Python v3.10.  This can
cause subtle failures, so this patch makes all three of these workflows
use Python v3.11.

This patch also fixes a bug in the RollPyTorch action so that it skips
creating a PR if the pytorch commit hash has not changed.