[MLIR][TORCH] Add E2E support for aten.embedding_dense_backward op
This commit adds the support for aten.embedding_dense_backward op by
decomposing it into a set of other torch ops.
This commit also removes this op from the list of decompositions used
by the dynamo compilation. Earlier, it was added to the list because
there was no lowering present for this op in Torch-MLIR.

Signed-Off By: Vivek Khandelwal <vivek@nod-labs.com>