[MLIR][TORCH] Add support for aten.tensor op
This commit reduces the `aten.tensor` op into the
`valsem.aten.tensor.<dtype>` which is lowered further.
This reduction is required since the aten.tensor op expects the
data list to be of any type `t` which is not a supported type in PyTorch.
Therefore, we reduce the op to the corresponding reduction op based on the dtype.

Signed-Off By: Vivek Khandelwal <vivek@nod-labs.com>