Layer_Normalization
Checking in 3 files for Layer_Normalization
There are limits to the implementation still:
onnx.Layer_Normalization produces 3 results while the
2nd and the 3rd is optional.
For this initial build we are ignoring those results since they
are not supported as part of the torch layer_norm.
Also added a function to check for the multiple result types in case we
would use them later.