Simplify lowering of aten.reflection_pad2d to linalg
The lowering uses a linalg.extract_slice operation to extract the
input slices to be used for the padding followed by a linalg.generic
operation to reverse the order of elements as required for the
reflection.

This commit removes the use of the linalg.generic. It uses the
capability of the linalg.extract_slice operation to extract a reversed
slice by specifying negative strides and offsets that point to the end
of the slice in the input tensor. This simplifies the generated IR
and makes it potentially more optimization friendly.