[torch-mlir][torch] add encoding field to torch type
This adds an encoding field to the torch type, using
the interfaces for printing, parsing, and verification.
Note that although this change prepares adding sparsity
to the torch type (as illustrated by the round trip and
invalid tests), nothing in this change depends on the
actual contents of the encoding field!