onnx: allow optional constant value for Pad operator
The `constant_value` input of the onnx.Pad operator is optional (see
https://onnx.ai/onnx/operators/onnx__Pad.html#inputs), but the existing
logic for lowering the operator into the Torch dialect assumes that it
is mandatory.

This patch makes the attribute optional and constructs a default value
(a list of zeros the size of the input tensor) if the attribute was not
specified.