onnx: fix preconditions for lowering AveragePool ops
The `pads` attribute of the AveragePool operator specifies the value to
pad at both the beginning as well as the end of the axis (see
https://onnx.ai/onnx/operators/onnx__AveragePool.html#attributes), so
the size of this attribute should be twice the rank of the input tensor.
However, our TorchOnnxToTorch bails out early since it incorrectly
compares the pads attribute with the rank (not twice the rank) of the
input tensor.

This patch fixes the code to match the spec and adds a lit test.