[torch-mlir][sparse] implement first sparse_jit end-to-end path
This PR introduces a sparse_jit wrapper that can run simple
models with sparse tensor inputs end-to-end. The implementation
shows all required components on modifying sparse tensor types
with a 1:N relation on the call sites. Two tests shows that the
JIT runs end-to-end while computing the correct results.

More details to follow (generalizing to COO and different ranks,
as well as support for *output* sparse tensors), but the general
concepts are all here now.