[fx] Implement auto_functionalized higher order op.
* Also adds the basic scaffolding for handling more of these, which will be needed for cond, while, etc.
* Refactors some of the support in the generic OpOverload emitter so it can be shared with these other special forms.

This has been on my list for a while, but it just so happens that as part of upgrading to PyTorch 2.3 and a pure upstream flow in Turbine, we were using a feature that required integration with auto_functionalized. This is perhaps the "weirdest" of the higher-order ops and a poor place to start, but needs must. We have testing for this in Turbine.

Full support in Turbine has an entire custom ops facility. I've reduced this down to a unit test in torch-mlir.