[MLIR][Torch] Canonicalize torch.from_i1 and torch.to_i1
When lowering torch.aten.convolution, it is expected that the 'transposed'
argument is a torch.constant operation. In some cases, the argument was a
from_i1 operation converting an arith.constant operation into a torch.bool.
This is not wrong semantically, but instead of generalizing the legality of the
torch.aten.convolution op, we canonicalize arith.constant ops followed by
from_i1 ops to torch.bool ops.

Additionally, when torch.to_i1 ops operate on a torch.constant : bool op,
we canonicalize those straight to arith.constant : i1 ops.