[fx] Support ExportedProgram buffer mutation.
In the prior state when I supported mutation of user inputs by treating them as mutable-tensor SSA values, I had left the case of buffer mutation only vaguely implemented until a concrete use emerged.

This patch reworks this buffer mutation support by assuming that buffers must be resolved via the hooks symbolically and treated with load/store semantics. This is implied in the structure since we have no SSA value that represents a buffer and we already assume that reading parameters happens via such a mechanism.